Rename all updates to patch
Remove service and used sharedouterapi client
