+++ src/ApprenticePortal/SFA.DAS.ApprenticePortal.Api/AppStart/AddServiceRegistrationExtension.cs
@@ -1,10 +1,6 @@
 ﻿using Microsoft.Extensions.DependencyInjection;
 using SFA.DAS.Api.Common.Infrastructure;
 using SFA.DAS.Api.Common.Interfaces;
-using SFA.DAS.ApprenticePortal.Application.Services;
-using SFA.DAS.ApprenticePortal.Clients;
-using SFA.DAS.ApprenticePortal.Configuration;
-using SFA.DAS.ApprenticePortal.Interfaces;
 using SFA.DAS.SharedOuterApi.Configuration;
 using SFA.DAS.SharedOuterApi.Infrastructure;
 using SFA.DAS.SharedOuterApi.Interfaces;
@@ -21,8 +17,6 @@ namespace SFA.DAS.ApprenticePortal.Api.AppStart
             services.AddTransient(typeof(IInternalApiClient<>), typeof(InternalApiClient<>));
             services.AddTransient<IApprenticeAccountsApiClient<ApprenticeAccountsApiConfiguration>, ApprenticeAccountsApiClient>();
             services.AddTransient<IApprenticeCommitmentsApiClient<ApprenticeCommitmentsApiConfiguration>, ApprenticeCommitmentsApiClient>();
-            services.AddTransient<IApprenticePortalApiClient<ApprenticePortalConfiguration>, ApprenticePortalApiClient>();
-            services.AddTransient<IApprenticePortalService, ApprenticePortalService>();
         }
     }
 }
