+++ src/ApprenticePortal/SFA.DAS.ApprenticePortal.Api/Controllers/ApprenticeController.cs
@@ -2,9 +2,8 @@
 using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
 using RestEase;
-using SFA.DAS.ApprenticePortal.Application.Commands.ApprenticeUpdate;
+using SFA.DAS.ApprenticePortal.Application.Commands.ApprenticePatch;
 using SFA.DAS.ApprenticePortal.Application.Queries.ApprenticeAccounts;
-using SFA.DAS.ApprenticePortal.InnerApi.ApprenticeAccounts.Requests;
 using SFA.DAS.ApprenticePortal.Models;
 using System;
 using System.Threading.Tasks;
@@ -37,7 +36,7 @@ namespace SFA.DAS.ApprenticePortal.Api.Controllers
         [HttpPatch("/apprentices/{apprenticeId}")]
         public async Task<IActionResult> UpdateApprentice([Path] Guid apprenticeId, [Body] JsonPatchDocument<Apprentice> patch)
         {
-            await _mediator.Send(new ApprenticeUpdateCommand
+            await _mediator.Send(new ApprenticePatchCommand
             {
                 ApprenticeId = apprenticeId,
                 Patch = patch
