+++ src/ApprenticePortal/SFA.DAS.ApprenticePortal/Application/Queries/ApprenticeAccounts/GetApprenticeQueryHandler.cs
@@ -1,10 +1,10 @@
 ﻿using MediatR;
-using System.Threading;
-using System.Threading.Tasks;
 using SFA.DAS.ApprenticePortal.InnerApi.ApprenticeAccounts.Requests;
 using SFA.DAS.ApprenticePortal.Models;
 using SFA.DAS.SharedOuterApi.Configuration;
 using SFA.DAS.SharedOuterApi.Interfaces;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace SFA.DAS.ApprenticePortal.Application.Queries.ApprenticeAccounts
 {
