--- src/ApprenticePortal/SFA.DAS.ApprenticePortal/Interfaces/IApprenticePortalApiClient.cs
@@ -1,9 +1,0 @@
-using SFA.DAS.SharedOuterApi.Interfaces;
-
-namespace SFA.DAS.ApprenticePortal.Interfaces
-{
-    public interface IApprenticePortalApiClient<T> : IInternalApiClient<T>
-    {
-        
-    }
-}
