--- src/ApprenticePortal/SFA.DAS.ApprenticePortal/Interfaces/ICommitmentsApiClient.cs
@@ -1,9 +1,0 @@
-using SFA.DAS.SharedOuterApi.Interfaces;
-
-namespace SFA.DAS.ApprenticePortal.Interfaces
-{
-    public interface ICommitmentsApiClient<T> : IInternalApiClient<T>
-    {
-        
-    }
-}
