+++ src/Shared/SFA.DAS.SharedOuterApi/Interfaces/IApprenticeAccountsApiClient.cs
@@ -1,6 +1,6 @@
 ﻿namespace SFA.DAS.SharedOuterApi.Interfaces
 {
-    public interface IApprenticeAccountsApiClient<T> : IGetApiClient<T>
+    public interface IApprenticeAccountsApiClient<T> : IApiClient<T>
     {
         
     }
