+++ src/Shared/SFA.DAS.SharedOuterApi/Services/ApprenticeAccountsApiClient.cs
@@ -1,4 +1,5 @@
-﻿using SFA.DAS.SharedOuterApi.Configuration;
+﻿using System.Collections.Generic;
+using SFA.DAS.SharedOuterApi.Configuration;
 using SFA.DAS.SharedOuterApi.Interfaces;
 using SFA.DAS.SharedOuterApi.Models;
 using System.Net;
@@ -28,6 +29,61 @@ namespace SFA.DAS.SharedOuterApi.Services
         public Task<ApiResponse<TResponse>> GetWithResponseCode<TResponse>(IGetApiRequest request)
         {
             return _apiClient.GetWithResponseCode<TResponse>(request);
+        }
+
+        public Task<IEnumerable<TResponse>> GetAll<TResponse>(IGetAllApiRequest request)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public Task<PagedResponse<TResponse>> GetPaged<TResponse>(IGetPagedApiRequest request)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public Task<TResponse> Post<TResponse>(IPostApiRequest request)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public Task Post<TData>(IPostApiRequest<TData> request)
+        {
+            return _apiClient.Post<TData>(request);
+        }
+
+        public Task Delete(IDeleteApiRequest request)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public Task Patch<TData>(IPatchApiRequest<TData> request)
+        {
+            return _apiClient.Patch<TData>(request);
+        }
+
+        public Task Put(IPutApiRequest request)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public Task Put<TData>(IPutApiRequest<TData> request)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public Task<ApiResponse<TResponse>> PostWithResponseCode<TResponse>(IPostApiRequest request, bool includeResponse = true)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public Task<ApiResponse<string>> PatchWithResponseCode<TData>(IPatchApiRequest<TData> request)
+        {
+            throw new System.NotImplementedException();
+        }
+
+        public Task<ApiResponse<TResponse>> PutWithResponseCode<TResponse>(IPutApiRequest request)
+        {
+            throw new System.NotImplementedException();
         }
     }
 }
